Add files via upload
the read and write enable signals are separated or both happen at the same time? Is there a time the fifo is full? the output form is accepted as block with 8bits store in it already? could it be like the block is updated over time and without all 8 bits at the beginning of a read cycle?