matlab generator script
this matlab script generates a random sequence of binary numbers, then prints those numbers to a document. also, this outputs a file that includes the address of dout in 13 bit binary as well as the actual output dout in order to compare to the test results in a .mif file.